Ignore public IPs when collecting metadata (#7)
Instances may have public IPs assigned after they are provisioned. The `vprox` server will fail to start if public IPs aren't yet assigned to the instance. This seems unnecessary because it doesn't need to know the public IP of the instance at all as far as I understand.

This change makes it easier to start vprox servers in ec2 instances with multiple ENIs but no public IPs, and then assign public IPs later. (AWS doesn't let you automatically assign public IPs if an instance is launched with multiple ENIs so there's a period in which we have no public IPs.)